UHF-13191: Add new computedSum field for sum fields
