+++ public/modules/custom/grants_application/fixtures/iakkaiden_kulttuuri_ja_liikunta/uiSchema.json
@@ -119,9 +119,11 @@
           "hideNameFromPrint": true
         },
         "applied_grant_amount_2026": {
-          "ui:field": "subventionSum",
+          "ui:field": "computedSum",
           "ui:options": {
-            "printableName": "applied_grant_amount_2026.title"
+            "printableName": "applied_grant_amount_2026.title",
+            "sourceFields": ["grant_info_step.subventions_section.subventions"],
+            "sourceShape": "subvention"
           }
         },
         "applied_grant_amount_2027": {
@@ -298,7 +300,11 @@
         "ui:widget": "hidden"
       },
       "applied_benefits": {
-        "ui:field": "subventionSum"
+        "ui:field": "computedSum",
+        "ui:options": {
+          "sourceFields": ["grant_info_step.subventions_section.subventions"],
+          "sourceShape": "subvention"
+        }
       },
       "income_rows": {
         "ui:options": {
