+++ public/modules/custom/grants_application/fixtures/kasvatus_ja_koulutus_toiminta_av/uiSchema.json
@@ -144,7 +144,11 @@
         "ui:widget": "hidden"
       },
       "applied_benefits": {
-        "ui:field": "subventionSum"
+        "ui:field": "computedSum",
+        "ui:options": {
+          "sourceFields": ["grant_info_step.subventions_section.subventions"],
+          "sourceShape": "subvention"
+        }
       }
     },
     "other_benefits_section": {
